delightex: float -> double

+++ jgltf-impl-v2/src/main/java/de/javagl/jgltf/impl/v2/CameraOrthographic.java
@@ -25,25 +25,25 @@ public class CameraOrthographic
      * The floating-point horizontal magnification of the view. (required) 
      * 
      */
-    private Float xmag;
+    private Double xmag;
     /**
      * The floating-point vertical magnification of the view. (required) 
      * 
      */
-    private Float ymag;
+    private Double ymag;
     /**
      * The floating-point distance to the far clipping plane. `zfar` must be 
      * greater than `znear`. (required)<br> 
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float zfar;
+    private Double zfar;
     /**
      * The floating-point distance to the near clipping plane. (required)<br> 
      * Minimum: 0.0 (inclusive) 
      * 
      */
-    private Float znear;
+    private Double znear;
 
     /**
      * The floating-point horizontal magnification of the view. (required) 
@@ -52,7 +52,7 @@ public class CameraOrthographic
      * @throws NullPointerException If the given value is <code>null</code>
      * 
      */
-    public void setXmag(Float xmag) {
+    public void setXmag(Double xmag) {
         if (xmag == null) {
             throw new NullPointerException((("Invalid value for xmag: "+ xmag)+", may not be null"));
         }
@@ -65,7 +65,7 @@ public class CameraOrthographic
      * @return The xmag
      * 
      */
-    public Float getXmag() {
+    public Double getXmag() {
         return this.xmag;
     }
 
@@ -76,7 +76,7 @@ public class CameraOrthographic
      * @throws NullPointerException If the given value is <code>null</code>
      * 
      */
-    public void setYmag(Float ymag) {
+    public void setYmag(Double ymag) {
         if (ymag == null) {
             throw new NullPointerException((("Invalid value for ymag: "+ ymag)+", may not be null"));
         }
@@ -89,7 +89,7 @@ public class CameraOrthographic
      * @return The ymag
      * 
      */
-    public Float getYmag() {
+    public Double getYmag() {
         return this.ymag;
     }
 
@@ -104,7 +104,7 @@ public class CameraOrthographic
      * the given constraints
      * 
      */
-    public void setZfar(Float zfar) {
+    public void setZfar(Double zfar) {
         if (zfar == null) {
             throw new NullPointerException((("Invalid value for zfar: "+ zfar)+", may not be null"));
         }
@@ -122,7 +122,7 @@ public class CameraOrthographic
      * @return The zfar
      * 
      */
-    public Float getZfar() {
+    public Double getZfar() {
         return this.zfar;
     }
 
@@ -136,7 +136,7 @@ public class CameraOrthographic
      * the given constraints
      * 
      */
-    public void setZnear(Float znear) {
+    public void setZnear(Double znear) {
         if (znear == null) {
             throw new NullPointerException((("Invalid value for znear: "+ znear)+", may not be null"));
         }
@@ -153,7 +153,7 @@ public class CameraOrthographic
      * @return The znear
      * 
      */
-    public Float getZnear() {
+    public Double getZnear() {
         return this.znear;
     }
