+++ jgltf-model/src/main/java/de/javagl/jgltf/model/v2/GltfModelCreatorV2.java
@@ -982,19 +982,19 @@ public class GltfModelCreatorV2
                 nodeModel.setCameraModel(cameraModel);
             }
             
-            float matrix[] = node.getMatrix();
-            float translation[] = node.getTranslation();
-            float rotation[] = node.getRotation();
-            float scale[] = node.getScale();
+            double matrix[] = node.getMatrix();
+            double translation[] = node.getTranslation();
+            double rotation[] = node.getRotation();
+            double scale[] = node.getScale();
             nodeModel.setMatrix(Optionals.clone(matrix));
             nodeModel.setTranslation(Optionals.clone(translation));
             nodeModel.setRotation(Optionals.clone(rotation));
             nodeModel.setScale(Optionals.clone(scale));
             
-            List<Float> weights = node.getWeights();
+            List<Double> weights = node.getWeights();
             if (weights != null)
             {
-                float weightsArray[] = new float[weights.size()];
+                double weightsArray[] = new double[weights.size()];
                 for (int j = 0; j < weights.size(); j++)
                 {
                     weightsArray[j] = weights.get(j);
@@ -1149,7 +1149,7 @@ public class GltfModelCreatorV2
             materialModel.setAlphaMode(AlphaMode.valueOf(alphaModeString));
         }
         materialModel.setAlphaCutoff(
-            Optionals.of(material.getAlphaCutoff(), 0.5f));
+            Optionals.of(material.getAlphaCutoff(), 0.5));
         
         materialModel.setDoubleSided(
             Boolean.TRUE.equals(material.isDoubleSided()));
@@ -1164,7 +1164,7 @@ public class GltfModelCreatorV2
             materialModel.setBaseColorTexcoord(
                 baseColorTextureInfo.getTexCoord());
         }
-        float[] baseColorFactor = Optionals.of(
+        double[] baseColorFactor = Optionals.of(
             pbrMetallicRoughness.getBaseColorFactor(),
             pbrMetallicRoughness.defaultBaseColorFactor());
         materialModel.setBaseColorFactor(baseColorFactor);
@@ -1179,12 +1179,12 @@ public class GltfModelCreatorV2
             materialModel.setMetallicRoughnessTexcoord(
                 metallicRoughnessTextureInfo.getTexCoord());
         }
-        float metallicFactor = Optionals.of(
+        double metallicFactor = Optionals.of(
             pbrMetallicRoughness.getMetallicFactor(),
             pbrMetallicRoughness.defaultMetallicFactor());
         materialModel.setMetallicFactor(metallicFactor);
         
-        float roughnessFactor = Optionals.of(
+        double roughnessFactor = Optionals.of(
             pbrMetallicRoughness.getRoughnessFactor(),
             pbrMetallicRoughness.defaultRoughnessFactor());
         materialModel.setRoughnessFactor(roughnessFactor);
@@ -1199,7 +1199,7 @@ public class GltfModelCreatorV2
             materialModel.setNormalTexcoord(
                 normalTextureInfo.getTexCoord());
             
-            float normalScale = Optionals.of(
+            double normalScale = Optionals.of(
                 normalTextureInfo.getScale(),
                 normalTextureInfo.defaultScale());
             materialModel.setNormalScale(normalScale);
@@ -1215,7 +1215,7 @@ public class GltfModelCreatorV2
             materialModel.setOcclusionTexcoord(
                 occlusionTextureInfo.getTexCoord());
             
-            float occlusionStrength = Optionals.of(
+            double occlusionStrength = Optionals.of(
                 occlusionTextureInfo.getStrength(),
                 occlusionTextureInfo.defaultStrength());
             materialModel.setOcclusionStrength(occlusionStrength);
@@ -1232,7 +1232,7 @@ public class GltfModelCreatorV2
                 emissiveTextureInfo.getTexCoord());
         }
         
-        float[] emissiveFactor = Optionals.of(
+        double[] emissiveFactor = Optionals.of(
             material.getEmissiveFactor(),
             material.defaultEmissiveFactor());
         materialModel.setEmissiveFactor(emissiveFactor);
